cleaning up the code and have output more succint

+++ InterpScript.py
@@ -10,26 +10,17 @@ from scipy.interpolate import InterpolatedUnivariateSpline
 
 NumToRun=1000
 
-#musDatamtn=pd.read_hdf('/n/holystore01/LABS/guenette_lab/Users/lrogers/Proposal/CombinedProposalOutputs.h5','Muons')
-#musDatamtn=pd.read_hdf('/n/holystore01/LABS/guenette_lab/Users/lrogers/Proposal/CombinedProposalInterpOutputs.h5','Muons')
+#pull in the combined proposal outputs
 musDatamtn=pd.read_hdf('/lcrc/project/NEXT/data/Muon-Rate-Measurement/Proposal/Outputs2022/CombinedProposalInterpOutputs.h5','MuonsProp')  
-#musDatamtn['perc']=musDatamtn.apply(lambda row: len(row.FinalMuons)/NumToRun, axis=1)
-#musDatamtn=musDatamtn[musDatamtn.X!=1295]
 
-#save interpolatations for proposal output for all X,Y for each energy
 
-#musDatamtn=musDatamtn[(musDatamtn.X>-600)&(musDatamtn.X<600)]
-#musDatamtn=musDatamtn[(musDatamtn.Y>-600)&(musDatamtn.Y<600)]
-#musDatamtn=musDatamtn[np.cos(np.deg2rad(musDatamtn.Theta))>=.4]
+#get interpolatations for proposal output for all X,Y for each energy
 
 
 energies=musDatamtn.Energy.unique()
 #print(energies)
 energies.sort()
-#print(len(energies))
 
-#energies=[5e5,6e5,7e5,8e5,9e5,10e5]
-#energies=[100,300,400,450,500,550,600,650,700,650,800,900,1000,2000,4000,6000] #in MeV  
 
 for LOC in range(0,len(energies)):
     r=0
@@ -81,4 +72,5 @@ for LOC in range(0,len(energies)):
     PercentMuons=scipy.interpolate.RectBivariateSpline(np.unique(df.X),np.unique(df.Y),percsarray,s=0,kx=3, ky=3)
     
 
-    with open('/lcrc/project/NEXT/data/Muon-Rate-Measurement/Proposal/Proposal_Muons_interpolator'+str(NRG*10**-3)+'GeV.pkl', 'wb') as f:                                          pickle.dump(PercentMuons, f)                            
+    with open('/lcrc/project/NEXT/data/Muon-Rate-Measurement/Proposal/Proposal_Muons_interpolator'+str(int(NRG*10**-3))+'GeV.pkl', 'wb') as f:
+        pickle.dump(PercentMuons, f)
